move cmap samples to utils
